Done the rebase.
Moved the function into the init of XMLEncoder.
