+++ implementation/mangaCheck/app/src/main/java/it/unitn/disi/lpsmt/g03/mangacheck/library/xml/XMLEncoder.kt
@@ -3,11 +3,14 @@ package it.unitn.disi.lpsmt.g03.mangacheck.library.xml
 import it.unitn.disi.lpsmt.g03.mangacheck.utils.xml.LibraryEntry
 import it.unitn.disi.lpsmt.g03.mangacheck.utils.xml.XMLEncoder
 import android.content.Context
+import android.util.Log
+import android.util.Xml
 import it.unitn.disi.lpsmt.g03.mangacheck.R
 import org.w3c.dom.Document
 import org.w3c.dom.Element
 import org.w3c.dom.NodeList
 import java.io.File
+import java.io.FileOutputStream
 import javax.xml.parsers.DocumentBuilderFactory
 import javax.xml.transform.TransformerFactory
 import javax.xml.transform.dom.DOMSource
@@ -15,6 +18,35 @@ import javax.xml.transform.stream.StreamResult
 
 
 class XMLEncoder(private val context: Context) : XMLEncoder<LibraryEntry> {
+
+    private val readingListFile = File(context.filesDir, context.getString(R.string.library_XML))
+
+    init{
+        if (!readingListFile.exists()) {
+
+            val outputFile: FileOutputStream =
+                context.openFileOutput(context.getString(R.string.library_XML), Context.MODE_PRIVATE)
+
+            val serializer = Xml.newSerializer()
+            serializer.setOutput(outputFile, "UTF-8")
+            serializer.startDocument("UTF-8", true)
+
+            serializer.startTag(null, "libraries")
+
+            serializer.endTag(null, "libraries")
+
+            serializer.endDocument()
+            serializer.flush()
+
+            outputFile.flush()
+            outputFile.close()
+        }
+        Log.v(
+            XMLEncoder::class.simpleName,
+            context.applicationContext!!.openFileInput(context.getString(R.string.library_XML)).bufferedReader()
+                .readText()
+        )
+    }
 
     override fun addEntry(entry: LibraryEntry){
         val builder = DocumentBuilderFactory.newInstance().newDocumentBuilder()
